Redesign `AtomicDurableOutputStream`
This class had the dangerous behavior that the file would be created
even if the code writing it threw an exception.

The new design requires an explicit call to `commit()` to make the
changes "real".  If the writer throws an exception, then commit is not
called and the file is not written.